Move all SwiftyStoreKit types into separate file
